requirements.txt
Flask==2.1.0
google-cloud-firestore==2.11.0
google-cloud-secret-manager==2.16.1
gunicorn==20.1.0
